+++ packages/backend/src/common/commonFormatAttributes.ts
@@ -18,7 +18,7 @@ export const formatStyleAttribute = (
 };
 
 export const formatDataAttribute = (label: string, value?: string) =>
-  ` data-${lowercaseFirstLetter(label)}${value === undefined ? `` : `="${value}"`}`;
+  ` data-${lowercaseFirstLetter(label).replace(" ", "-")}${value === undefined ? `` : `="${value}"`}`;
 
 export const formatClassAttribute = (
   classes: string[],
