+++ packages/backend/src/index.ts
@@ -3,5 +3,4 @@ export { htmlMain } from "./html/htmlMain";
 export { tailwindMain } from "./tailwind/tailwindMain";
 export { swiftuiMain } from "./swiftui/swiftuiMain";
 export { run } from "./code";
-export { convertNodesToAltNodes as convertIntoNodes } from "./altNodes/altConversion";
 export * from "./messaging";
