+++ packages/plugin-ui/src/components/ColorsPanel.tsx
@@ -14,14 +14,13 @@ const ColorsPanel = (props: {
   };
 
   return (
-    <div className="bg-gray-100 dark:bg-neutral-900 w-full rounded-lg p-4 flex flex-col gap-2">
-      <div className="p-0 pb-2 border-neutral-200 dark:border-neutral-700">
+    <div className="bg-card border w-full rounded-lg p-4 flex flex-col gap-2">
+      <div className="p-0 pb-2">
         <div className="flex items-center justify-between">
-          <h2 className="text-lg font-semibold text-neutral-800 dark:text-neutral-100 flex items-center gap-2">
-            {/* <div className="w-4 h-4 rounded-full bg-gradient-to-br from-blue-400 to-blue-600"></div> */}
+          <h2 className="text-lg font-semibold text-foreground flex items-center gap-2">
             Color Palette
           </h2>
-          <span className="text-xs bg-neutral-100 dark:bg-neutral-700 px-2 py-1 rounded-full text-neutral-500 dark:text-neutral-400">
+          <span className="text-xs bg-muted dark:bg-muted px-2 py-1 rounded-xl text-muted-foreground">
             {props.colors.length} color{props.colors.length > 1 ? "s" : ""}
           </span>
         </div>
@@ -33,7 +32,7 @@ const ColorsPanel = (props: {
             key={"button" + idx}
             className={`w-full h-16 rounded-lg text-sm font-semibold shadow-sm transition-all duration-300 ${
               isPressed === idx
-                ? "ring-4 ring-green-300 ring-opacity-50 animate-pulse"
+                ? "ring-4 ring-primary ring-opacity-50 animate-pulse"
                 : "ring-0"
             }`}
             style={{ backgroundColor: color.hex }}
